requirements.txt
pillow
requests
requests[socks]
aiohttp